Fixed circular dependencies for checks
